Split base Spec into partials by spec writing methods vs test case generation
